requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pip==21.3.1
pydantic-core==2.14.5
pydantic==2.5.2
requests==2.31.0
setuptools==60.2.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
urllib3==2.1.0
websocket-client==1.6.4
websockets==12.0
wheel==0.37.1
yarl==1.9.3
